+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/application/Exceptions.kt
@@ -1,6 +1,7 @@
 package eu.tib.orkg.prototype.statements.application
 
 import eu.tib.orkg.prototype.contributions.domain.model.ContributorId
+import eu.tib.orkg.prototype.statements.domain.model.ClassId
 import eu.tib.orkg.prototype.statements.domain.model.ObservatoryId
 import eu.tib.orkg.prototype.statements.domain.model.OrganizationId
 import eu.tib.orkg.prototype.statements.domain.model.ResourceId
@@ -55,6 +56,10 @@ class ClassNotAllowed(`class`: String) : RuntimeException("This class id ($`clas
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 class ClassAlreadyExists(`class`: String) : RuntimeException("The class with the id ($`class`) already exists")
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+class InvalidClassCollection(ids: Iterable<ClassId>) :
+    RuntimeException("The collection of classes $ids contains one or more missing classes")
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 class DuplicateURI(uri: URI, id: String) :