+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- When the resource's classes are updated, all classes need to exist. (Closes: [#216](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/216))
 ### Added
 - The feature flag `orkg.features.use-neo4j-version-3` was introduced to skip initialization steps that are not compatible with Neo4j 4.x.